new corrections

+++ 3.1.3/src/main/java/ru/kata/spring/boot_security/demo/repositories/RoleRepository.java
@@ -2,12 +2,12 @@ package ru.kata.spring.boot_security.demo.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.kata.spring.boot_security.demo.models.Role;
+import ru.kata.spring.boot_security.demo.models.entity.Role;
 
 import java.util.List;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     List<Role> findAll();
     Role getRoleByName(String name);
 }
